+++ surveyform/src/components/form/FormLayout.tsx
@@ -1,4 +1,3 @@
-"use client";
 import SurveySectionHeading from "~/components/questions/SurveySectionHeading";
 import FormNav from "./FormNav";
 import { FormElement } from "./FormElement";
@@ -8,29 +7,18 @@ import { ReactNode } from "react";
 import { SectionMetadata } from "@devographics/types";
 import ReadingList from "../reading_list/ReadingList";
 import FormSectionMessage from "./FormSectionMessage";
-import FormMessages from "./FormMessages";
 
-export interface FormLayoutProps
-  extends Omit<
-    FormInputProps,
-    "path" | "value" | "question" | "questionNumber"
-  > {
+export interface FormLayoutProps {
+  section: SectionMetadata;
+  enableReadingList?: boolean;
   children: ReactNode;
-  sectionNumber: number;
   nextSection: SectionMetadata;
   previousSection: SectionMetadata;
 }
 
 export const FormLayout = (props: FormLayoutProps) => {
-  const {
-    children,
-    section,
-    sectionNumber,
-    edition,
-    nextSection,
-    previousSection,
-    enableReadingList,
-  } = props;
+  const { children, section, nextSection, previousSection, enableReadingList } =
+    props;
 
   return (
     <div className="survey-layout">
