+++ surveyform/src/components/form/typings.ts
@@ -1,20 +1,13 @@
 import type { ResponseDocument } from "@devographics/types";
 import {
-  SurveyMetadata,
-  EditionMetadata,
   // OptionMetadata,
   // DbPaths,
-  SectionMetadata,
   QuestionMetadata,
 } from "@devographics/types";
+import { FormProps } from "./FormPropsContext";
 // import { Message } from "./FormMessages";
 
 export type FormInputProps<TValue = string | number | string[] | number[]> = {
-  /**
-   * NOTE: in read-only mode there might be no response
-   * All form inputs have to be robust to this scenario
-   */
-  response?: ResponseDocument;
   /**
    * 
    * NOTE: "path" may be undefined if the component template name (for instance "textList")
@@ -22,15 +15,12 @@ export type FormInputProps<TValue = string | number | string[] | number[]> = {
    */
   path: string;
   value: TValue; // TODO: value might be undefined?
-  survey: SurveyMetadata;
-  edition: EditionMetadata;
-  section: SectionMetadata;
   question: QuestionMetadata;
-  updateCurrentValues: any;
-  submitForm: any;
-  readOnly?: boolean;
-  stateStuff: any;
-  sectionNumber: number;
   questionNumber: number;
   enableReadingList: boolean;
-};
+  updateCurrentValues: any
+  /** @deprecated Get from context instead */
+  response?: ResponseDocument
+}
+  // courtesy to avoid rewriting all input components
+  & FormProps;
